+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2023-08-15
 
 - 实现 `wakatime` 插件基础功能 | Implement the basic functions of the `wakatime` plugin.
+- 调整打开 `siyuan://plugins` 链接事件的名称 | Adjust the name of the event to open the `siyuan://plugins` link.
 
 ## 2023-08-14